perf(evm): widen the stack head to native width
Head is read and modified by every push and pop, so it is the most frequent
narrow access left in the interpreter. The zkEVM guest bills a 4-byte
read-modify-write at 122 to read and 193 to write, against 16 and 18 for an
aligned 8-byte one.

It costs no space: CodeLength moving to native width left four bytes of
padding after Head, and widening it fills them, so EvmStack stays 40 bytes.

On the guest, block 25,532,382: MEMORY -5.79%, OPCODES -0.92%, total cost
-0.63%. Memory now sits 3.09% below the merge base rather than above it, and
the dispatch chain is at +1.23% overall, from +1.88%.

The locals that carry Head through the pop paths widen with it, so the
computation no longer narrows and then widens again to index the stack.

x64 pays a little for this: +99 bytes over the 151 instantiations the opcode
benchmark compiles, 27 of them larger by at most 10 bytes and none smaller,
against a 0.63% saving on the target that charges for the access.

+++ src/Nethermind/Nethermind.Evm/VirtualMachine.cs
@@ -1337,7 +1337,7 @@ public partial class VirtualMachine<TGasPolicy>(
         {
             if (TTracingInst.IsActive)
                 EndInstructionTrace(TGasPolicy.GetRemainingGas(in gas));
-            UpdateCurrentState((int)programCounter, in gas, stack.Head);
+            UpdateCurrentState((int)programCounter, in gas, (int)stack.Head);
         }
         else
         {
@@ -1410,7 +1410,7 @@ public partial class VirtualMachine<TGasPolicy>(
 
         if (_txTracer.IsTracingStack)
         {
-            _txTracer.SetOperationStack(new TraceStack(vmState.MemoryStacks(stackValue.Head)));
+            _txTracer.SetOperationStack(new TraceStack(vmState.MemoryStacks((int)stackValue.Head)));
         }
 
         if (_txTracer.IsTracingReturnData)
